backend/src/auth_api/authenticate.py
from django.conf import settings

from rest_framework_simplejwt.authentication import JWTAuthentication
from rest_framework import authentication, exceptions as rest_exceptions


def enforce_csrf(request):
    check = authentication.CSRFCheck(request)
    reason = check.process_view(request, None, (), {})
    if reason:
        raise rest_exceptions.PermissionDenied("CSRF Failed: %s" % reason)


class CustomAuthentication(JWTAuthentication):
    def authenticate(self, request):
        raw_token = request.COOKIES.get(settings.SIMPLE_JWT["AUTH_COOKIE"])

        if raw_token is None:
            header = self.get_header(request)
            if header is None:
                return None

            raw_token = self.get_raw_token(header)
            if raw_token is None:
                return None

        try:
            validated_token = self.get_validated_token(raw_token)
        except:
            return None

        if raw_token == request.COOKIES.get(settings.SIMPLE_JWT["AUTH_COOKIE"]):
            enforce_csrf(request)

        return self.get_user(validated_token), validated_token
